# subscription-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://www.cloudbees.com/products/terms-service",
        "contact": {
            "name": "CloudBees Support",
            "url": "http://support.cloudbees.com",
            "email": "support@cloudbees.com"
        },
        "license": {}
    },
    "host": "localhost:8085",
    "basePath": "/api/v1",
    "paths": {
        "/accounts": {
            "get": {
                "description": "Gets an array of accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetAccounts",
                "operationId": "cloud-bill-saas-subscription-service-get-accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "optional comma separated list of filter",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "optional order",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/persistence.Account"
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Upsert an account passing account json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upsert an account",
                "operationId": "cloud-bill-saas-subscription-service-upsert-account",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/persistence.Account"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Upserted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/accounts/{accountId}": {
            "get": {
                "description": "Retrieves an account by account ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get an account",
                "operationId": "cloud-bill-saas-subscription-service-get-account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/persistence.Account"
                        }
                    },
                    "400": {
                        "description": "Missing account ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an account",
                "operationId": "cloud-bill-saas-subscription-service-delete-account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing account ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/contacts": {
            "get": {
                "description": "Gets an array of contacts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetContacts",
                "operationId": "cloud-bill-saas-subscription-service-get-contacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "optional comma separated list of filter",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "optional order",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/persistence.Contact"
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Upsert a contact passing contact json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upsert a contact",
                "operationId": "cloud-bill-saas-subscription-service-upsert-contact",
                "responses": {
                    "204": {
                        "description": "Upserted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contacts/{accountId}": {
            "get": {
                "description": "Retrieves an contact by account ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get an contact",
                "operationId": "cloud-bill-saas-subscription-service-get-contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/persistence.Contact"
                        }
                    },
                    "400": {
                        "description": "Missing account ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an contact",
                "operationId": "cloud-bill-saas-subscription-service-delete-contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing account ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/entitlements": {
            "get": {
                "description": "Gets an array of entitlements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetEntitlements",
                "operationId": "cloud-bill-saas-subscription-service-get-entitlements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "optional comma separated list of filter",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "optional order",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/persistence.Entitlement"
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Upsert an entitlement passing entitlement json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upsert an entitlement",
                "operationId": "cloud-bill-saas-subscription-service-upsert-entitlement",
                "responses": {
                    "204": {
                        "description": "Upserted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/entitlements/{entitlementId}": {
            "get": {
                "description": "Retrieves an entitlement by entitlement ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get an entitlement",
                "operationId": "cloud-bill-saas-subscription-service-get-entitlement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entitlement ID",
                        "name": "entitlementId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/persistence.Entitlement"
                        }
                    },
                    "400": {
                        "description": "Missing entitlement ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an entitlement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an entitlement",
                "operationId": "cloud-bill-saas-subscription-service-delete-entitlement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entitlement ID",
                        "name": "entitlementId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing entitlement ID in path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "persistence.Account": {
            "type": "object",
            "properties": {
                "approvals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/persistence.Approval"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "persistence.Approval": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "persistence.Contact": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "emailAddress": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    }
}